new system based on boxes
Every element is a box with a specified width and height. These can be
infinite (rendered with a scrollbar). They use the specified width and
height but constrain them to the dimensions of their parent, and then
render with padding. The dimensions leftover after padding are given to
the children, which each constrain themselves and return a view. These
views are then rendered with specified spacing between them left to
right and top to bottom, with no centering.